cleanup

+++ nextjs/app/Portfolio/components/SpecificPortfolio.tsx
@@ -1,4 +1,4 @@
-import { Box, Heading, Button, Flex } from "@chakra-ui/react";
+import { Box, Heading, Button, Flex, Text } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import {
   useDeletePortfolioMutation,
@@ -12,16 +12,26 @@ export default function SpecificPortfolio() {
   const [deletePortfolio] = useDeletePortfolioMutation({
     refetchQueries: ["GetPortfolios"],
   });
-  const { data } = useGetPortfolioQuery({
+  const { data, loading, error } = useGetPortfolioQuery({
     variables: { portfolioID },
     fetchPolicy: "cache-and-network",
   });
   const portfolio = data?.portfolios_by_pk;
+
+  if (!data && loading) return "Loading...";
+  console.log({ data, error });
+
   return (
     <Box pt={16}>
       <Flex justifyContent="flex-end" mb={2}>
         <Button
           onClick={() => {
+            if (portfolio?.members?.length) {
+              const sure = confirm(
+                `If you delete this portfolio, ${portfolio?.members?.length} donors will be unassigned. Are you sure?`
+              );
+              if (!sure) return;
+            }
             deletePortfolio({ variables: { portfolioID } });
             router.push("/");
           }}
@@ -32,14 +42,16 @@ export default function SpecificPortfolio() {
       </Flex>
       <Heading>{portfolio?.name}</Heading>
 
+      {!portfolio?.members?.length && <Text pt={8}>No members yet</Text>}
+
       <PersonList
-        people={portfolio?.members.map(({ donor }) => donor)}
-        Actions={({ infoID }: any) => {
+        people={portfolio?.members.map(({ crm }) => crm)}
+        Actions={({ momentumID }: any) => {
           return (
             <Button
               colorScheme="blue"
               onClick={() => {
-                router.push(`/portfolio/${portfolioID}/donor/${infoID}`);
+                router.push(`/portfolio/${portfolioID}/donor/${momentumID}`);
               }}
             >
               View
